Update ------
This project is behind schedule, In my bid to meet up with the given deadline
I rushed the project and failed to effectively commit the changes.

With this commit, I am pushing this project based on what I've accomplished so far.

The database model has been created, the APIs are under way.